+++ developer-knowledgebase/shared-node-data.md
@@ -9,7 +9,6 @@ Any data saved this way is technically publicly accessible. More about the share
 | `updatedAt`            | `string`                           | This is property saved the last updated timestamp and it's only saved on `figma.root`                                                                                                                  |
 | `values`               | `Record<string, AnyTokenSet>`      | This is key stores the actual token values (if remote storage is not enabled). This is also saved on `figma.root`                                                                                      |
 | `storageType`          | `StorageType`                      | This key saves the currently selected (remote) storage provider. This value will never include the API secret of the provider                                                                          |
-| `hash`                 | `string`                           | This is the checksum of the current token object for a single node. This is used to keep track of whether nodes need to be updated or not                                                              |
 | `usedTokenSet`         | `UsedTokenSetsMap`                 | This is the map used to determine the selected token sets. This used to be a `string[]` array but is being migrated to a map                                                                           |
 | `activeTheme`          | `string`                           | This is used to keep track of which theme was activated                                                                                                                                                |
 | `themes`               | `ThemeObjectsList`                 | This key saves all the available themes. (only if remote storage is not enabled)                                                                                                                       |
