+++ src/utils/formatNumber.ts
@@ -1,0 +1,8 @@
+export function formatNumber(num: number): string {
+  if (num >= 1000000) {
+    return num.toString();
+  } if (num >= 1000) {
+    return `${(num / 1000).toFixed(1)}k`;
+  }
+  return num.toString();
+}
